Adding the pytz 2015.4 modulze for timezone support.
pytz can be found at https://pypi.python.org/pypi/pytz/.

The downloaded package was pyt2-2015.4-py2.7.egg and I've extracted it
and placed it directly in the repository.

Eventually I should probably turn this script into a package and wire
this dependency properly but for now it will have to suffice as a direct
copy so that a simple git clone works out of the box.