Fixed the initial commit error.
In one of my previous changes I made the commit look up if there are any
previous commits in order to ensure only changes are added and
everything of interest is reported.

This generated an error for the initial commit since no commit existed.

This patch adds a flag for the initial commit which prevents the commit
method from looking for a previous commit.