Delete empty directories before performing an accurev pop.
All empty directories will be recreated by the accurev pop command if
deleted. They are inexpensive to remove and repopulate but should have
theoretically appeared in our accurev diff. If it is inexpensive it's
better to be safe then sorry. I have no time to verify what accurev does
right now so best to be pesimistic.