Changed the format and content of commit notes.
The script state is tracked for each stream in a note that is attached
to each commit (in a namespace that has the same name as the branch).
This state is intended to tag each commit and allow you to uniquely
identify from which transaction it originated. The minimum information
required is the depot, transaction number and the name of the stream
which we were processing when we generated this commit.