Fixing up the git filter-branch command format.
I've abstracted out a fair amount of the clutter (copy/paste code) from the
StitchBranches function.

Removed a bug where the branch name ended up in the aliasMap.

The resulting merge map is not ideal and could be further worked on but
is sufficient to achieve merge requests between the branches.