No-op transaction error message is now debug output.
The error message which notifies you that a transaction wasn't committed
has proven to be redundant. Occasionally its output can be large so I've
condensed it to a single line and moved additional information to the
debug output of the program which can be shown by using the -v switch on
the commandline.

This is a particular case and should be treated with caution, which is
why I'm still leaving a one-line message for the user. However, in 100%
of the times I've seen this message it related to the way accurev did
versioning and paranoid output of the accurev diff command.