Symlinks on a Linux host need to be deleted with os.remove().
It seems that the symbolic links in Python 2.7
only work correctly on Linux hosts. Windows symlinks
are not correctly recongnized in 2.7.

Additionally, to delete a symlink we can't use
os.rmdir() and must use os.remove(). This patches
the issue but there is a broader question if the
item the symlink points to also needs to be deleted
or not. Currently only the symlink is deleted.