Changed temporary files to use tempfile library.
Local temporary files used to store the commit message and the notes
used to use hard-coded names which means that they will potentially
conflict if two instances of the script run at the same time.

This commit resolves this for the commit message and the notes which are
the only culprits to my knowledge. If they are not it should be trivial
to apply this change whereever else it is needed.