Improving error handling when switching branches.
This commit introduces a few exceptions for situations that were
previously left ambiguous. The script should fail if it gets invalid
data back from the commandline, be it from git or accurev.

Additionally, it is expensive to simply delete everything, every time.
Though the paranoid approach of deleting everything is beneficial, and
was intentionally introduced so that you would be guaranteed to
"re-synchronize" with the AccuRev content at least every time you switch
branches, it is too expensive now that the converted git repository has
been validated to be correct.

This speeds up the branch switches by approximately 15 seconds in my
tests.