Enforcing invariants via exceptions.
It is important that the assumptions made through out the program hold.
On occasion I've seen strange errors that can arrise from invalid state
returned by various commands.

This commit attempts to address this by throwing exceptions and causing
the conversion process to fail when a command returns an obviously
incorrect result.