Preparing the stage for a new conversion method.
It is possible to iterate over the entire history and make good
decisions about where a git merge should occur. However, this requires
an entirely different approach and this commit paves the way for it to
be implemented along-side the existing methods.

Since this work is controlled via a switch it should not yet affect the
existing code so it is ok to include in the master branch.