Fixing initial state assumptions for process streams.
The state checks for the return values of the 'git branch' and 'git
status' commands were not correct and have caused the script to be
unable to detect the scenario where one of them returned an invalid
result.

When this happened, the processing of a stream would start from the
beginning, even though there is existing history in the repo. This
would be seen as a normal git history which at one point goes back in
time to the beginning and repeats itself.

The checks included in this commit should mitigate this issue.