Fixing the issue with unknown source streams.
It is not always possible to use `accurev hist -fex` to determine the
source stream of a transaction. In these cases the best we can do is
cherry-pick for now. It should be possible to figure out more state
later via more complex algorithms but for now the best we can do is a
cherry-pick.