Fixing initial state assumptions.
It seams that unless we are on a branch git shows us to be in a detached
HEAD state even if we are on a hidden ref's commit chain. Additionally,
the `git status` command shows us to be detached from the first commit
in the hidden ref's history and refers to the first commit hash rather
than the ref name in the output.

My initial state assumption checks rely on the output of `git status`
giving me the branch/ref name. Since this is no longer the case the
easiest way to verify that I'm in fact on the branch I expect to be on
is to verify that the HEAD commit hash matches the hidden ref's commit
hash.