Implemented the data processing draft.
The processing of the accurev information which is stored in the
refs/ac2git/<depot-name>/streams/stream_<stream-number>_info ref history
is now complete and stores 3 files in each commit:
  - hist.xml
  - streams.xml
  - diff.xml

Note: the diff.xml file is omitted for the mkstream transactions.

For each stream we first get all of the above mentioned information for
the transactions that affect the stream and then do an `accurev pop` for
each of these on a separate ref,
refs/ac2git/<depot-name>/streams/stream_<stream-number>_data.

Each commit on either ref has a one line message with the following
format: 'transaction <transaction number>' which is then used as a key
into this data store or accurev state and stream contents.

The idea is to retrieve information about the transactions affecting a
stream into these two refs, meaning that we import the contents of the
stream, the output of the `accurev hist` command, the output of the
`accurev show streams` command and the output of the `accurev diff -a -i
-v <streamName> -V <streamName> -fx` command with which we can make
decisions about merging the streams at a later date.

Once everything is downloaded from accurev we will reprocess these two
refs transaction by transaction, probably in conjunction with other
streams' pairs of refs, and produce a merged output.

The expensive operation is the downloading from Accurev so adding a
stream later should be a lot cheaper, even if we have to redo the
merging, since we will already have all the necessary information.