Removing the finalize option.
The finalize option was useful in the beginning but will be superceeded
by the feature being developed in this branch.

The idea is that choosing a merge strategy would implicitly do the post
processing of streams now that the retrieval of accurev information has
been decoupled from the processing of the branches. The new way of
post-processing is a lot cheaper and less destructive.