Adding hidden ref for depots and moving streams under depotNumber instead of depotName.
In the case where we started processing a depot and it is renamed
during processing the script will have to restart if we encode the
refs with the depot name. Hence we should encode it with the depot
number instead.

This commit achieves that by maintaining a refs/ac2git/depots branch
which has the history of a single depots.xml file that is the result
of the output of the `accurev show depots -fix` command.

If the depot name requested, from the config file, is not found in
the depots.xml then we execute the accurev command and commit a new
depots.xml (only if it has changed).

The refs/ac2git/depots ref has potential merge issues when you're
using the script in a distributed manner (which is entirely possible
and intended). It will require manual merging and there's no way
around it since the `accurev show depots` command only shows you
the depots at the current point in time and is unable to provide
historical information.