Addresses issue 42
The GetStreamMap() function was just plainly wrong. This should now
return something sensible for when no stream list is specified.

The example config was also fixed. @yfw008 pointed out that the `--`
token is invalid in an XML comment string and I used it in an example
git command (mainly `git log --notes=accurev`) which is useful to keep.
As a compromise I've removed one of the hyphens in the hope that the
user that actually reads the comment will realise this small mistake
and be able to correct it by reading the `git help log` manual.