Addresses issue #42.
The problem is that when we have an empty streams list we won't
retrieve the full stream-list from accurev because we were trying
to avoid adding extra calls to the accurev server. However, this
is not necessary and it is acceptable that if the user wishes to
be in `disconnected` mode w.r.t. accurev server that he is required
to have a specific stream list in the config.