Fixes issue #39. Folders with `..` are now checked when deleting.
The problem was that someone promoted a folder named `\.\Folder\..` to Accurev,
which in turn caused my script to try and delete the contents of the worktree.
Sicne this particular function didn't check that what it is deleting is or is not
the .git directory the consequence was a destroyed conversion. Sepuku if you will.

This should now be fixed since I've added a check to see if we are deleting the
git repository folder itself to the DeleteDiffItemsFromRepo() which then in turn
calls the ClearGitRepo() for this case which does understand that it shouldn't
delete the .git directory so all is well... For now.