Fix the merging end state #45
Having the "next transaction" to process saved in the state, though
useful is not correct. It caused the script to continuously reprocess
the same "next transaction" over and over again once it completed.

In order to fix this I've changed the script so that it stores the
"last transaction" which was processed instead.