First step of fixing issue #57. Adding state info.
Added more hidden state information for the merging
stage that should allow me to better determine the
commit hashes for each branch at any point in history.

Since merging makes determinig where each branch was
inconvenient I will maintain hidden refs for each
stream that will contain a commit tree that is linear
when traversed via the first parent:

  `git log --first-parent`

The first commit on this hidden ref has no parents
(i.e. orphaned branch) and every subsequent commit
has two parents. The first parent is the previous
commit on the hidden ref and the second parent is
the commit that was made on this branch, while the
commit message denotes for which transaction this
commit occurred.

With this we can use this hidden ref like this:

  `git log --first-parent --before=<some-date> \
     --format=format:%P`

To get the commit hash of that particular stream
at this particular point in time (which corresponds
directly to a transaction number).