Addresses GitHub issue #64: Restart conversion...
Restart conversion is broken and needs more options

This issue is not entirely accurate. Currently a
request to restart a conversion deletes the entire
repository directory and starts again.

This is highly destructive and since the process
has been split into two phases we can be a little
smarter about what we are redoing. It shouldn't
be necessary to re-download all Accurev information
but it would be useful to be able to reprocess the
converted repository to pickup a new stream that
may have been added _(warning: doing so only makes
sense before the repository is shared with others
or made live, don't do this on already published
conversions, monitor GitHub issue #59 for a fix for
that situation)_.