Simplifying ProcessStream() function.
Related to GitHub issue #59.

The ProcessStream() function deals with cherry picking
individual transactions onto an orphaned branch in git.

This function now uses the same helpers that the
ProcessTransaction() function uses, most notably the
CommitTransaction() wrapper instead of dealing with
the Commit() function directly.

Consequently the state that ProcessStream()
generates and uses in order to be able to continue from
where it stopped last time now matches ProcessTransaction().
This will allow us to use ProcessStream() on branches that
are added to the configuration file after a conversion has
already started in order to bring them up to the current
high water mark after which ProcessTransaction() can continue
with the merging process.

All that remains is to add the ability for us to tell
ProcessStream() to create the branch at a particular commit
if it doesn't exist already which should be trivial at this
point i.e. Whenever the `parents` list is empty in the
ProcessStream() function we need to set it to

`[ startAtThisCommitHash ]`

and we are done.