Draft implementation of add/rename/remove streams.
Addresses GitHub issue #59.

This patch implements a rudimentary strategy for
managing additions, removals and name changes to
the configuration file w.r.t. branches that have
already been processed by the merge algorithm.

For removed branches a hidden reference is left
behind so that if this stream is readded it can
resume from where it was last left.

Renaming branches is naive and fails if the
new branch name already exists.

Adding a new branch will simply cherry pick all
transactions from its creation up to the last
transaction that was processed with the merging
algorithm after which the merging algorithm will
continue processing with the new branch included.

Added branches are always orphaned and this
should be addressed in the next patch.