Adding ability to attach the added stream to its basis.
I've expanded the ability of ProcessStream() to encompass
additions of streams that are attached to their basis
streams. ProcessStream() doesn't have any responsibility
towards sequencing streams and this should fall onto
the caller to ensure that the basis exists for the child
stream. Since this isn't implemented in this patch an
assert will trigger when a basis stream has an existing
branch but we fail to find a point at which we can attach
our new branch. This assert is too paranoid since it is
possible that the basis stream simply doesn't have a
commit at the given time (i.e. we set a timelock that is
way earlier than the basis creation time) and that we
would be justified in making an orphan branch but it
is best to alert the user to the things that the algorithm
doesn't handle correctly than silently fail so lets be
overprotective.