Fix the handling of 'files that break history'.
The `git diff` command has a number of flags that may or may not
respect the recommended way of excluding files that break history.

Most notably we use the `git diff --stat` command since it gives us
a smaller output but it turns out that this command doesn't respect
the `.git/info/attributes` as noted in this StackOverflow comment:

http://stackoverflow.com/questions/10415100/want-to-exclude-file-from-git-diff#comment29471399_10421385

Hence we should revert back to using the plain `git diff` command.

Additionally, the GitDiff() wrapper function already ensures that
the returned diff is not None and it also strips the result of
any whitespace so there is no need to do this again on the caller
side.