Fix the branch point for streams added after the fact
Once a conversion or tracking process has been started it is
possible to add streams which will be processed on the next
iteration of the script. These streams need to determine a branch
point and this bug caused the incorrect branch point to be used.

To fix it the stream.basisStreamNumber needs to represent the basis
stream number at the time the stream was created which may or may
not be the same as the stream.basisStreamNumber that the stream is
currently a child of.