Fix incorrect branch point of new streams
If a stream is added to an already running conversion as the case
would be with a tracked repository, the script didn't correctly
account for the passage of time and would simply attach the new
stream to the tip of its basis.

This was due to the code for addition of new streams being
retrofitted to existing code and incorrectly using functions with
specific underlying assumptions.

The GetBasisCommitHash() function assumed that it was being called
at the time the stream was created and that the basis stream could
not be _ahead_ in time. This holds true if we are incrementally
adding transactions but doesn't hold when we are attaching a new
stream to a conversion that is already in progress.

The code has been patched and GetBasisCommitHash() has been given
an extra parameter that needs to be specified so that it knows
the expected stream creation time. This time is compared with the
timelock that is specified on the new stream to ensure that we
pick the earliest time at which we will attach the new stream's
git branch.

This new parameter can be specified as None in which case the
existing behavior will be used.