Updated ignored transaction types list
Not all Accurev transactions are of interest to the conversion as
some Accurev transactions deal with server maintenance or issue
linkage and do not physically affect the stream's default set (file
contents).

Hence, these should be ignored by the script when processing a
depot.

A hard-coded list of transactions has been added to the top of the
script that specifies the transaction types that are not of interest.
The list is based on a few test cases but since documentation of the
transactions in question is not readily available it is possible that
the transactions should in fact not be ignored. Hence, this list
will be printed each time the script is invoked to remind the user of
this decision.

Resolves: issue #70