Add filtering by stream type of auto added streams
When a stream list isn't specified the user can now specify what types
of streams are to be excluded from the automatically tracked streams.

Most commonly this will be used to exclude workspaces from being
converted.

Hence the feature is simple and only available via the config file.
To use, simply add a comma separated list of stream types to the
exclude-types attribute of the stream-list element in the config (see
example configuration comments for details).

Resolves: issue #71
# Please enter the commit message for your changes. Lines starting
# with '#' will be kept; you may remove them yourself if you want to.
# An empty message aborts the commit.
# On branch master
# Your branch is up-to-date with 'origin/master'.
#
# Changes to be committed:
#	modified:   ac2git.py
#