Fixup author-is-commiter option
The author-is-committer option used to raise an exception when set
to false which was left over from the legacy algorithm that relied
on the dates in the generated branches to make decisions.

With the introduction of commit_history refs this was no longer the
case but the exception remeained.

The hidden refs all force the author-is-committer option to "true"
since they are used as the data backend but the generated branches
now obey the option correctly and produce the desired result.