Refactor GetStreamByName to use accurev
The second stage of processing does its utmost to avoid calling any
accurev commands. However, this results in a slow pick-axe search
through the downloaded history for the old names of renamed streams.

This patch checks if we are logged into Accurev and uses the accurev
command to get the stream number from its name only resorting to the
previous mechanism if Accurev is actually unavailable.