Issue #60: Infer source stream
Implemented new algorithm to infer the source stream of an Accurev
promote transaction. This is done when this information is not available
in the `accurev hist -fevx` output, which is the case for Accurev
transactions recorded with old versions of accurev and affects depots
which have gone through several upgrades.

This new algorithm is experimental and is controlled via flags which
default to being disabled and clearly state that it is an experimental
feature.