Allow missing mkstream transactions
When the first transaction in a stream is not a `mkstream` transaction the
script now takes the first transaction that it can find for the given stream.
This was not tested sufficiently, but does resolve a the issue 100 which was
reported on Github. Evidently, transaction 1 can have multiple streams in
existence at the same time, and is not a `mkstream` transaction.